feat@refactor internals

+++ source/_consumeTypings.test.js
@@ -1,3 +1,4 @@
+import isCI from 'is-ci'
 import { resolve } from 'path'
 
 import { execCommand } from '../files/execCommand.js'
@@ -7,6 +8,8 @@ jest.setTimeout(3 * 60 * 1000)
 const DIR = resolve(__dirname, '../../')
 
 test('typings can be imported', async () => {
+  if (!isCI) return
+
   await execCommand('rm -rf rambda-scripts-clone', DIR)
   await execCommand('yarn build:main')
